requirements.txt
aiofiles
Pillow
humanize
hachoir
motor
pyrofork==2.3.45
TgCrypto
dnspython
aiohttp
pyromod==1.5
telethon
instaloader
tiktok_downloader
tiktok-scraper
shazamio
spotipy
TikTokApi

yt-dlp>=2023.3.4
